// Frontend/cijene-frontend/src/types/api.ts
export interface LoginRequest {
  email: string;
  password: string;
  rememberMe?: boolean;
}